requirements.txt
grpcio==1.75.1
protobuf==6.32.1
